+++ CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * `futhark lsp` provides inlay hints, they show type ascriptions for
     inferred types of bindings, by VegOwOtenks. (#2398)
 
+* Scan chunk size is now exposed as tuning parameter in `cuda` and `hip`
+  backends.
+
 ### Removed
 
 ### Changed